:sparkles: feat(ui): 다시 시작하기 버튼 클릭에 따른 게임 초기화 추가
