+++ src/js/app.js
@@ -1,4 +1,5 @@
 import Store from './flux/store';
+import initialState from './flux/initialState';
 
 import './components/MoneyForm';
 import './components/LottoList';
@@ -8,19 +9,6 @@ import './components/WinningNumberForm';
 import './components/ResultModal';
 
 import '../css/index.css';
-
-const initialState = {
-  money: 0,
-  lottoList: [],
-  lottoListVisibility: false,
-  winningNumbers: ['', '', '', '', '', ''],
-  bonusNumber: 0,
-  resultModalVisibility: false,
-  result: {
-    winningCounts: { fifth: 0, fourth: 0, third: 0, second: 0, first: 0 },
-    earningsRate: 0,
-  },
-};
 
 window.store = new Store(initialState);
