+++ src/js/flux/actions.js
@@ -2,6 +2,7 @@ const ACTION = {
   PURCHASE_LOTTO: 'purchase-lotto',
   TOGGLE_LOTTO_LIST: 'toggle-lotto-list',
   TOGGLE_RESULT_MODAL: 'toggle-result-modal',
+  RESTART: 'restart',
 };
 
 export default ACTION;
